bot/requirements.txt
addict==2.4.0
diffusers==0.24.0
accelerate==0.25.0
fiftyone==0.23.1
gradio==4.8.0
huggingface_hub==0.19.4
ipdb==0.13.13
matplotlib==3.8.2
numpy==1.26.2
opencv_python==4.8.1.78
Pillow==10.1.0
psycopg2_binary==2.9.9
pyTelegramBotAPI==4.14.0
PyYAML==6.0.1
Requests==2.31.0
setuptools==68.0.0
termcolor==2.4.0
timm==0.9.12
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
typer==0.9.0
yapf==0.40.2
